# backend/swagger.yaml
openapi: 3.0.3
info:
  title: MyActivities API
  description: |
    API REST de la plateforme **MyActivities** — découverte, planification et évaluation d'activités locales.

    ## Authentification
    L'API utilise des **JWT Bearer tokens** (access token 15 min + refresh token 7 jours).
    Les endpoints protégés nécessitent le header : `Authorization: Bearer <access_token>`

    ## Rôles utilisateurs
    | Rôle | Description |
    |------|-------------|
    | `visitor` | Non connecté — lecture seule |
    | `member` | Compte créé — planning, avis |
    | `organizer` | Référence ses activités |
    | `admin` | Modération et administration |

    ## Règles métier importantes
    - **1 avis par utilisateur par activité** (enforced côté backend)
    - **Toute nouvelle activité requiert une validation admin** avant publication
    - **Brute force** : 5 tentatives échouées → blocage 15 min
    - **Géolocalisation** : consentement explicite requis (RGPD Art. 6.1.a)
    - **Suppression de compte** : efface toutes les données personnelles (RGPD Art. 17)

    ## Codes d'erreur communs
    | Code | Signification |
    |------|---------------|
    | 400 | Données invalides |
    | 401 | Token manquant ou expiré |
    | 403 | Accès refusé (rôle insuffisant) |
    | 404 | Ressource introuvable |
    | 409 | Conflit (ex : email déjà utilisé) |
    | 429 | Trop de requêtes (rate limiting) |

  version: 1.0.0
  contact:
    name: Charles Deschamps
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (expiration 15 min)

  # ─────────────────────────────────────────────
  # SCHEMAS
  # ─────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 3

    # ── Errors ──────────────────────────────────
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Le champ email est invalide.
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # ── Auth ────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: john.doe@example.com
        pseudo:
          type: string
          example: JohnDoe42
        role:
          type: string
          enum: [member, organizer, admin]
          example: member
        status:
          type: string
          enum: [active, suspended, deleted]
          example: active
        siret:
          type: string
          nullable: true
          description: SIRET de l'organisateur (14 chiffres) ; null pour un membre.
          example: "73282932000074"
        avatarUrl:
          type: string
          format: uri
          nullable: true
          example: https://cdn.myactivities.app/avatars/uuid.jpg
        createdAt:
          type: string
          format: date-time
        emailVerified:
          type: boolean
          example: true
...
    ActivityCategory:
      type: string
      enum: [sport, culture, gastronomie, nature, bien_etre, art, musique, autre]
    ActivityStatus:
      type: string
      enum: [pending, published, rejected, archived]
    ActivitySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Escalade en salle — ClimbUp Lyon
        category:
          $ref: '#/components/schemas/ActivityCategory'
        distance:
          type: number
          format: float
          description: Distance en km depuis la position de l'utilisateur
          example: 1.4
        avgRating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.2
        reviewCount:
          type: integer
          example: 37
        priceMin:
          type: number
          format: float
          example: 8.0
        priceMax:
          type: number
          format: float
          example: 15.0
        coverImage:
          type: string
          format: uri
          nullable: true
        address:
          type: string
          example: 12 Rue de la République, 69001 Lyon
        latitude:
          type: number
          format: float
          example: 45.7640
        longitude:
          type: number
          format: float
          example: 4.8357
    ActivityDetail:
      allOf:
        - $ref: '#/components/schemas/ActivitySummary'
        - type: object
          properties:
            description:
              type: string
              example: Salle d'escalade indoor avec murs de 12m et blocs tous niveaux.
            openingHours:
              type: object
              description: Horaires par jour de la semaine
              example:
                monday: "10:00-22:00"
                tuesday: "10:00-22:00"
                wednesday: "10:00-22:00"
                thursday: "10:00-22:00"
                friday: "10:00-23:00"
                saturday: "09:00-22:00"
                sunday: "09:00-20:00"
            accessibility:
              type: object
              properties:
                pmr:
                  type: boolean
                  description: Accessible PMR
                  example: true
                stroller:
                  type: boolean
                  description: Accessible poussette
                  example: false
            photos:
              type: array
              items:
                type: string
                format: uri
            websiteUrl:
              type: string
              format: uri
              nullable: true
            status:
              $ref: '#/components/schemas/ActivityStatus'
            adminNote:
              type: string
              nullable: true
              description: Motif de rejet renseigné par l'admin (null sinon)
            organizer:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                pseudo:
                  type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateActivityRequest:
      type: object
      required: [name, category, description, address, latitude, longitude, priceMin, priceMax]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          example: Escalade en salle — ClimbUp Lyon
        category:
          $ref: '#/components/schemas/ActivityCategory'
        description:
          type: string
          minLength: 20
          maxLength: 2000
        address:
          type: string
          example: 12 Rue de la République, 69001 Lyon
        latitude:
          type: number
          format: float
          example: 45.7640
        longitude:
          type: number
          format: float
          example: 4.8357
        priceMin:
          type: number
          format: float
          minimum: 0
          example: 8.0
        priceMax:
          type: number
          format: float
          minimum: 0
          example: 15.0
        openingHours:
          type: object
          example:
            monday: "10:00-22:00"
            friday: "10:00-23:00"
        accessibility:
          type: object
          properties:
            pmr:
              type: boolean
            stroller:
              type: boolean
        websiteUrl:
          type: string
          format: uri
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri
          maxItems: 10
    UpdateActivityRequest:
      type: object
      description: Tous les champs sont optionnels — PATCH sémantique
      properties:
        name:
          type: string
        category:
          $ref: '#/components/schemas/ActivityCategory'
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        priceMin:
          type: number
          format: float
        priceMax:
          type: number
          format: float
        openingHours:
          type: object
        accessibility:
          type: object
          properties:
            pmr:
              type: boolean
            stroller:
              type: boolean
        websiteUrl:
          type: string
          format: uri
          nullable: true
        photos:
          type: array
          items:
            type: string
            format: uri

    # ── Planning ─────────────────────────────────
...
    AdminUser:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            suspendedUntil:
              type: string
              format: date-time
              nullable: true
            reviewCount:
              type: integer
              example: 5
            planningCount:
              type: integer
              example: 12
...
  parameters:
    activityId:
      name: activityId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

  # ─────────────────────────────────────────────
  # RESPONSES
  # ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Token manquant ou expiré
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Token JWT manquant ou expiré.
    Forbidden:
      description: Rôle insuffisant pour cette action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: Accès réservé aux administrateurs.
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: Ressource introuvable.
...
paths:
...
  /activities:
    get:
      tags: [Activities]
      summary: Lister les activités géolocalisées
      description: |
        Retourne les activités autour d'une position GPS avec filtres optionnels.
        Accessible sans authentification (visiteur).
        La géolocalisation de l'utilisateur doit être transmise par le client (avec consentement RGPD explicite).
      operationId: listActivities
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: float
          description: Latitude de l'utilisateur
          example: 45.7640
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: float
          description: Longitude de l'utilisateur
          example: 4.8357
        - name: radius
          in: query
          schema:
            type: number
            format: float
            default: 5
            minimum: 0.5
            maximum: 50
          description: Rayon de recherche en km (défaut 5 km)
        - name: category
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ActivityCategory'
          style: form
          explode: false
          description: Filtrer par catégories (choix multiple)
        - name: priceMin
          in: query
          schema:
            type: number
            format: float
            minimum: 0
          description: Prix minimum (€)
        - name: priceMax
          in: query
          schema:
            type: number
            format: float
          description: Prix maximum (€)
        - name: minRating
          in: query
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 5
          description: Note moyenne minimale
        - name: search
          in: query
          schema:
            type: string
            minLength: 3
          description: Recherche textuelle (nom, description, catégorie)
        - name: sort
          in: query
          schema:
            type: string
            enum: [distance, rating, price_asc, price_desc]
            default: distance
          description: Critère de tri
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Liste paginée d'activités
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivitySummary'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          description: Paramètres lat/lng manquants ou invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Activities]
      summary: Créer une activité
      description: |
        Réservé aux organisateurs.
        L'activité est créée avec le statut `pending` et doit être validée par un admin avant publication (BF-037).
      operationId: createActivity
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Activité créée — en attente de validation admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Rôle organizer requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /activities/{activityId}:
    get:
      tags: [Activities]
      summary: Obtenir la fiche d'une activité
      description: Retourne tous les détails d'une activité publiée. Accessible sans authentification.
      operationId: getActivity
      parameters:
        - $ref: '#/components/parameters/activityId'
      responses:
        '200':
          description: Fiche activité complète
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Activities]
      summary: Modifier une activité
      description: |
        Réservé à l'organisateur propriétaire ou à un admin.
        Une modification remet l'activité en statut `pending` si des champs clés sont modifiés.
      operationId: updateActivity
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityRequest'
      responses:
        '200':
          description: Activité mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Activities]
      summary: Supprimer une activité
      description: Réservé à l'organisateur propriétaire ou à un admin.
      operationId: deleteActivity
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
      responses:
        '204':
          description: Activité supprimée
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  # PLANNING
  # ══════════════════════════════
...
  /organizers/me/activities:
    get:
      tags: [Organizers]
      summary: Lister ses activités
      description: Retourne toutes les activités de l'organisateur connecté (tous statuts inclus).
      operationId: getOrganizerActivities
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ActivityStatus'
          description: Filtrer par statut
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Activités de l'organisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityDetail'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Rôle organizer requis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Lister les utilisateurs
      description: Réservé aux administrateurs. Retourne tous les comptes avec leur statut.
      operationId: adminListUsers
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, deleted]
        - name: role
          in: query
          schema:
            type: string
            enum: [member, organizer, admin]
        - name: search
          in: query
          schema:
            type: string
          description: Recherche par pseudo ou email
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Liste des utilisateurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}/suspend:
    patch:
      tags: [Admin]
      summary: Suspendre un compte utilisateur
      description: Suspension temporaire. Les actions de l'admin sont journalisées.
      operationId: adminSuspendUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, suspendUntil]
              properties:
                reason:
                  type: string
                  maxLength: 500
                  example: Comportement abusif répété dans les avis.
                suspendUntil:
                  type: string
                  format: date-time
                  description: Date de fin de suspension
      responses:
        '200':
          description: Compte suspendu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}:
    delete:
      tags: [Admin]
      summary: Supprimer un compte utilisateur
      description: Suppression définitive par l'admin. Action journalisée avec motif.
      operationId: adminDeleteUser
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  maxLength: 500
      responses:
        '204':
          description: Compte supprimé
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/role:
    patch:
      tags: [Admin]
      summary: Changer le rôle d'un utilisateur
      description: Promotion ou rétrogradation (member ↔ organizer ↔ admin). Action journalisée.
      operationId: adminChangeUserRole
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [member, organizer, admin]
      responses:
        '200':
          description: Rôle mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/activities/pending:
    get:
      tags: [Admin]
      summary: Lister les activités en attente de validation
      operationId: adminListPendingActivities
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Activités en attente de validation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActivityDetail'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/activities/{activityId}/approve:
    patch:
      tags: [Admin]
      summary: Valider une activité
      description: |
        Passe l'activité de `pending` à `published`.
        L'organisateur est notifié de la publication.
      operationId: adminApproveActivity
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
      responses:
        '200':
          description: Activité publiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/activities/{activityId}/reject:
    patch:
      tags: [Admin]
      summary: Rejeter une activité
      description: |
        Passe l'activité de `pending` à `rejected`.
        L'organisateur est notifié avec le motif du rejet.
      operationId: adminRejectActivity
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/activityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  maxLength: 500
                  example: Description insuffisante et adresse non vérifiable.
      responses:
        '200':
          description: Activité rejetée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ══════════════════════════════
  # MESSAGING
  # ══════════════════════════════
...